refactor: 코드 컨벤션 수정
- struct 파일로 나눔
- 기본 값 제거
- 중복 기능 융합
