+++ OpenMarket/OpenMarket/Controllers/ProductsViewController.swift
@@ -126,7 +126,7 @@ extension ProductsViewController {
     }
     
     private func configureHierarchy() {
-        collectionView = UICollectionView(frame: view.bounds, collectionViewLayout: createListLayout())
+        collectionView = UICollectionView(frame: view.bounds, collectionViewLayout: createLayout(title: Titles.list))
         collectionView?.backgroundColor = .white
         collectionView?.delegate = self
         
@@ -176,29 +176,22 @@ extension ProductsViewController {
     }
     
     private func setLayout(_ collectionView: UICollectionView, _ topCellIndexPath: IndexPath) {
+        
+        guard let selectedSegmentIndex = self.segmentControl?.selectedSegmentIndex,
+              let title = Titles(rawValue: selectedSegmentIndex) else { return }
+        
+        self.isFetchingEnd = false
+        
         UIView.animate(withDuration: 0.3) { [weak self] in
             guard let self = self else { return }
-            if self.segmentControl?.selectedSegmentIndex == Titles.list.rawValue {
-                self.isFetchingEnd = false
-                collectionView.setCollectionViewLayout(self.createListLayout(), animated: false) { [weak self] bool in
-                    self?.isFetchingEnd = true
-                }
-                collectionView.visibleCells.forEach { cell in
-                    guard let cell = cell as? ItemCollectionViewCell else { return }
-                    cell.setAxis(segment: .list)
-                }
-                collectionView.scrollToItem(at: topCellIndexPath, at: .top, animated: false)
-            } else if self.segmentControl?.selectedSegmentIndex == Titles.grid.rawValue {
-                self.isFetchingEnd = false
-                collectionView.setCollectionViewLayout(self.createGridLayout(), animated: false) { [weak self] bool in
-                    self?.isFetchingEnd = true
-                }
-                collectionView.visibleCells.forEach { cell in
-                    guard let cell = cell as? ItemCollectionViewCell else { return }
-                    cell.setAxis(segment: .grid)
-                }
-                collectionView.scrollToItem(at: topCellIndexPath, at: .top, animated: false)
+            collectionView.setCollectionViewLayout(self.createLayout(title: title), animated: false) { bool in
+                self.isFetchingEnd = true
             }
+            collectionView.visibleCells.forEach { cell in
+                guard let cell = cell as? ItemCollectionViewCell else { return }
+                cell.setAxis(segment: title)
+            }
+            collectionView.scrollToItem(at: topCellIndexPath, at: .top, animated: false)
         }
     }
     
@@ -222,36 +215,37 @@ extension ProductsViewController {
 // MARK: - Layout Method
 
 extension ProductsViewController {
-    private func createListLayout() -> UICollectionViewLayout {
-        let flowLayout = UICollectionViewFlowLayout()
-        
-        let width = view.frame.width
-        let height = view.frame.height
-        
-        flowLayout.minimumLineSpacing = 2
-        flowLayout.estimatedItemSize = CGSize(width: width, height: height * 0.08)
-        
-        return flowLayout
-    }
-    
-    private func createGridLayout() -> UICollectionViewLayout {
-        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1.0),
-                                              heightDimension: .fractionalHeight(1.0))
-        let item = NSCollectionLayoutItem(layoutSize: itemSize)
-        
-        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1.0),
-                                               heightDimension: .fractionalHeight(0.31))
-        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize,
-                                                       subitem: item,
-                                                       count: 2)
-        group.interItemSpacing = .fixed(10)
-        
-        let section = NSCollectionLayoutSection(group: group)
-        section.interGroupSpacing = 10
-        section.contentInsets = NSDirectionalEdgeInsets(top: 10, leading: 10, bottom: 10, trailing: 10)
-        
-        let layout = UICollectionViewCompositionalLayout(section: section)
-        return layout
+    private func createLayout(title: Titles) -> UICollectionViewLayout {
+        switch title {
+        case .list:
+            let flowLayout = UICollectionViewFlowLayout()
+            
+            let width = view.frame.width
+            let height = view.frame.height
+            
+            flowLayout.minimumLineSpacing = 2
+            flowLayout.estimatedItemSize = CGSize(width: width, height: height * 0.08)
+            
+            return flowLayout
+        case .grid:
+            let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1.0),
+                                                  heightDimension: .fractionalHeight(1.0))
+            let item = NSCollectionLayoutItem(layoutSize: itemSize)
+            
+            let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1.0),
+                                                   heightDimension: .fractionalHeight(0.31))
+            let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize,
+                                                           subitem: item,
+                                                           count: 2)
+            group.interItemSpacing = .fixed(10)
+            
+            let section = NSCollectionLayoutSection(group: group)
+            section.interGroupSpacing = 10
+            section.contentInsets = NSDirectionalEdgeInsets(top: 10, leading: 10, bottom: 10, trailing: 10)
+            
+            let layout = UICollectionViewCompositionalLayout(section: section)
+            return layout
+        }
     }
     
     private func addIndicatorLayout() {
